refactor(oidc): Move error types to the error module
